requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
scikit-learn==1.5.2
numpy==1.26.4
openai==1.31.0
python-dotenv==1.0.0
pydantic==2.5.0
chromadb==0.4.22
aiohttp==3.9.1